% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_mat_luv.R
\name{as_mat_luv}
\alias{as_mat_luv}
\title{Convert to LUV matrix from HCL data frame}
\usage{
as_mat_luv(df_hcl)
}
\arguments{
\item{df_hcl}{\code{data.frame} with variables \code{h}, \code{c}, \code{l}}
}
\value{
\code{matrix} with columns, \code{l}, \code{u}, \code{v} and same number
of rows as \code{df_hcl}
}
\description{
It is thought to specify a colorpath using HCL coordinates;
it seems reasonable to compile these coordinates using a
data frame.
}
\examples{
  as_mat_luv(df_hcl_blues)
}
