% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-classes.R
\name{colorpath_S3_classes}
\alias{colorpath_S3_classes}
\title{S3 classes}
\description{
This package uses S3 classes to help clarify expectations for
function arguments and return values.
}
\details{
\strong{\code{cpath_palette_luv}}

A function that:
\itemize{
\item takes a vector numeric argument \code{x}, each value expected to be
between 0 and 1
\item returns a matrix with a row for each value of \code{x},
and three columns: \code{l}, \code{u}, \code{v}
}

Philosophically, this acts just like a continuous-palette function,
except that \code{LUV} values are returned instead of hex-codes.

\strong{\code{cpath_rescaler}}

A function that:
\itemize{
\item takes a numerical input \code{x} (vector, each member between 0 and 1)
\item returns a rescaled numerical vector (call it \code{y}),
each member between 0 and 1
}

There shall be a monotonic relationship between \code{x} and \code{y}.
}
