% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{data_hcl}
\alias{data_hcl}
\title{Get HCL data}
\usage{
data_hcl(pal_luv, n = 11)
}
\arguments{
\item{pal_luv}{\code{function} with S3 class \code{"cpath_pal_luv"}}

\item{n, }{\code{integer} number of points to generate for the palette}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}} with variables \code{type}, \code{hex}, \code{hue}, \code{chroma},
\code{luminance}.
}
\description{
This is in support of plots to show chroma-luminance plane and
hue-chroma plane.
}
