% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_hl.R
\name{data_surface_hl}
\alias{data_surface_hl}
\title{Create HCL data frame for surface}
\usage{
data_surface_hl(sfc, step = 0.5)
}
\arguments{
\item{sfc}{\code{function} created using \code{surface_hl()}.}

\item{step}{\code{numeric}, step size in hue, luminance}
}
\value{
\code{tibble} with columns \code{h}, \code{c}, \code{l}, \code{hex}.
}
\description{
Create HCL data frame for surface
}
\examples{
  # multi hue
  sfc_blues_multi <- surface_hl(c(240, 260))
  df_sfc <- data_surface_hl(sfc_blues_multi) # lots of rows!
}
