% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_hl.R
\name{df_hcl}
\alias{df_hcl}
\title{Create HCL data-frame using HL surface}
\usage{
df_hcl(df_cl, sfc)
}
\arguments{
\item{df_cl}{\code{data.frame} with columns \code{c}, \code{l}.}

\item{sfc}{\code{function} created using \code{surface_hl()}.}
}
\value{
\code{tibble} with columns \code{h}, \code{c}, \code{l}
}
\description{
Create HCL data-frame using HL surface
}
\examples{
  sfc_blues_multi <- surface_hl(c(240, 260))
  df_cl <- data.frame(l = c(20, 50, 80), c = c(0, 150, 0))

  df_hcl(df_cl, sfc_blues_multi)
}
