% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_hcl_blues}
\alias{df_hcl_blues}
\title{HCL data frame}
\format{
\code{data.frame} with 3 rows and 3 variables:
\describe{
\item{h}{\code{numeric}, hue}
\item{c}{\code{numeric}, chroma}
\item{l}{\code{numeric}, luminance}
}
}
\usage{
df_hcl_blues
}
\description{
A dataset containing values describing colors along a sample colorpath, using
the HCL colorspace. Describes the same colors as \link{mat_luv_blues}.
}
\keyword{datasets}
