% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{get_distance}
\alias{get_distance}
\title{Get perceptual distance}
\usage{
get_distance(pal_luv, n = 20, method = "cie2000", ...)
}
\arguments{
\item{pal_luv}{\code{function} with S3 class \code{"cpath_pal_luv"}}

\item{n}{\code{numeric} number of intervals into which to split domain
of the palette.}

\item{method}{\code{character} method to use for distance, one of
\code{"cie2000"}, \code{"euclidean"}, \code{"cie1976"}, \code{"cie94"}, or \code{"cmc"}.}

\item{...}{other arguments sent to \code{\link[farver:compare_colour]{farver::compare_colour()}}.}
}
\value{
\code{numeric} vector, length \code{n}, perceptual distances for
each interval.
}
\description{
Uses \code{\link[farver:compare_colour]{farver::compare_colour()}} to calculate perceptual distances
along intervals over a palette-function.
}
