% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mat_luv_blues}
\alias{mat_luv_blues}
\title{LUV matrix}
\format{
\code{matrix} with 3 rows and 3 variables:
\describe{
\item{l}{\code{numeric}, luminance}
\item{u}{\code{numeric}, u component}
\item{v}{\code{numeric}, v component}
}
}
\usage{
mat_luv_blues
}
\description{
A dataset containing values describing colors along a sample colorpath, using
the LUV colorspace.  Describes the same colors as \link{df_hcl_blues}.
}
\keyword{datasets}
