% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_luv_bezier.R
\name{pal_luv_bezier}
\alias{pal_luv_bezier}
\title{Create palette-function using Bézier spline}
\usage{
pal_luv_bezier(mat_luv, rescale_path = TRUE, n = 21)
}
\arguments{
\item{mat_luv}{\code{matrix} of LUV coordinates, control-points for Bézier spline.}

\item{rescale_path}{\code{logical} indicating to rescale the palette function to be
perceptually-uniform with respect to the input.}

\item{n}{\code{numeric} if \code{rescale_path}, number of equally-spaced Bézier points to calculate.}
}
\value{
A function with S3 class \code{cpath_palette}.
}
\description{
This takes a \code{matrix} of LUV control-points and returns a palette-function
(that returns LUV values) based on a Bézier spline using those
control-points.
}
\details{
The default is to rescale the palette function so that it is perceptually
uniform (in LUV space) using 21 points from the spline.
}
\examples{
 # create palette-function
 pb <- pal_luv_bezier(mat_luv_blues)

 # evaluate
 pb(c(0, 0.5, 1))

}
