% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-bezier.R
\name{palette_bezier}
\alias{palette_bezier}
\title{Create palette-function using Bézier curves}
\usage{
palette_bezier(luv)
}
\arguments{
\item{luv}{\code{matrix} of LUV coordinates, control-points for Bézier spline.}
}
\value{
A function with S3 class \code{cpath_palette}.
}
\description{
Create palette-function using Bézier curves
}
