% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cl}
\alias{plot_cl}
\title{Plot chroma-luminance plane}
\usage{
plot_cl(pal_luv, n = 11, label_hue = FALSE)
}
\arguments{
\item{pal_luv}{\code{function} with S3 class \code{"cpath_pal_luv"}}

\item{n}{\code{integer} number of points to generate for the palette}

\item{label_hue}{\code{logical} indicates to add a text layer showing the
hue each control point.}
}
\value{
Object with S3 class \code{gg}.
}
\description{
Plot chroma-luminance plane
}
