% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_hl.R
\name{plot_surface_hl}
\alias{plot_surface_hl}
\title{Plot surface}
\usage{
plot_surface_hl(sfc, step = 0.5)
}
\arguments{
\item{sfc}{\code{function} created using \code{surface_hl()}.}

\item{step}{\code{numeric}, step size in hue, luminance}
}
\value{
Object with S3 class \code{gg}.
}
\description{
Plot surface
}
\examples{
  # multi hue
  sfc_blues_multi <- surface_hl(c(240, 260))
  plot_surface_hl(sfc_blues_multi)

}
