% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{pth_as_hex}
\alias{pth_as_hex}
\title{Coerce to hex code}
\usage{
pth_as_hex(hex)
}
\arguments{
\item{hex}{\code{character} that can be coerced to a six-digit hex code.}
}
\value{
\code{character}, six-character hex code.
}
\description{
Use to validate and coerce to character-based hex-codes.
We coerce to CSS six-digit hex codes that use lower-case,
e.g. \code{"#336699"}, \code{"#aabbcc"}.
}
\examples{
  pth_as_hex("#11223344")
  pth_as_hex("#AA1122")
}
