% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroma.R
\name{pth_clip_chroma}
\alias{pth_clip_chroma}
\title{Modify color to bring inside RGB gamut}
\usage{
pth_clip_chroma(color, ...)
}
\arguments{
\item{color}{\code{character} or \code{matrix} with S3 class \code{pth_mat}, a
representation of a color.}

\item{...}{other args (not used).}
}
\value{
Object of same type as \code{color}.
}
\description{
Modify color to bring inside RGB gamut
}
