% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-sfclumsat.R
\name{pth_color_sfclumsat}
\alias{pth_color_sfclumsat}
\title{Get color using surface, luminance, and saturation}
\usage{
pth_color_sfclumsat(surface, lum, sat = 1)
}
\arguments{
\item{surface}{object with S3 class \code{pth_surface}.}

\item{lum}{\code{double} luminance of the color, \verb{0 <= lum <= 100}.}

\item{sat}{\code{double} saturation of the color, \verb{0 <= sat <= 1}.}
}
\value{
Object with S3 class \code{pth_mat}, describing a color.
}
\description{
Get color using surface, luminance, and saturation
}
\examples{
  sfc_blue <- pth_new_surface("#0000FF")
  color <- pth_color_sfclumsat(sfc_blue, lum = 50, sat = 1)

  pth_to_hex(color)
}
