% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{pth_colorspace_name}
\alias{pth_colorspace_name}
\title{Get colorspace name}
\usage{
pth_colorspace_name(x, ...)
}
\arguments{
\item{x}{\code{object} for which we want the colorspace name.
Can be a color matrix, surface, or palette.}

\item{...}{other args, not used.}
}
\value{
\code{character} name of the colorspace, e.g. \code{"CIELUV"}
}
\description{
Get colorspace name
}
