% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam02ucs.R, R/cam16ucs.R, R/cielab.R,
%   R/cieluv.R, R/creator.R, R/jzazbz100.R
\name{pth_creator.pth_cam02ucs}
\alias{pth_creator.pth_cam02ucs}
\alias{pth_creator.pth_cam16ucs}
\alias{pth_creator.pth_cielab}
\alias{pth_creator.pth_cieluv}
\alias{pth_creator}
\alias{pth_creator.pth_jzazbz100}
\title{Extract creator function}
\usage{
\method{pth_creator}{pth_cam02ucs}(mat, ...)

\method{pth_creator}{pth_cam16ucs}(mat, ...)

\method{pth_creator}{pth_cielab}(mat, ...)

\method{pth_creator}{pth_cieluv}(mat, ...)

pth_creator(mat, ...)

\method{pth_creator}{pth_jzazbz100}(mat, ...)
}
\arguments{
\item{mat}{matrix with object with S3 class \code{pth_mat}}

\item{...}{other args, not used}
}
\value{
\code{function} that encapsulates a creator, i.e. \code{pth_new_cieluv}
}
\description{
Extract creator function
}
