% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-palette.R
\name{pth_data_cat_euclid}
\alias{pth_data_cat_euclid}
\alias{pth_data_cat_metric}
\alias{pth_data_qnt_euclid}
\alias{pth_data_qnt_metric}
\title{Get color distances within palette}
\usage{
pth_data_cat_euclid(color, ...)

pth_data_cat_metric(color, ...)

pth_data_qnt_euclid(palette, ...)

pth_data_qnt_metric(palette, ...)
}
\arguments{
\item{color}{Object that can be coerced into colors,
i.e. \code{pth_hex} or \code{pth_mat}. Used for categorical palettes.}

\item{...}{additional arguments passed on to \code{transformer}.}

\item{palette}{\code{function} with S3 class \code{pth_palette}. Used for quantitative
palettes.}
}
\value{
\code{tbl_df} with columns:
\describe{
\item{hex_original_a}{\code{character} hex-code of original color.}
\item{hex_original_b}{\code{character} hex-code of original color.}
\item{cvd}{\code{character} type of color-vision deficiency.}
\item{severity}{\code{numeric} indicates severity using scale 0-1.}
\item{hex_a}{\code{character} hex-code of color under CVD.}
\item{hex_b}{\code{character} hex-code of color under CVD.}
\item{distance}{\code{numeric} perceptual distance between colors}
}
Each row describes an interaction between two colors in a palette.
For a categorical palette  (\verb{_cat} functions), each color in the palette is
compared with all the colors in the palette; for \code{n} colors, a tibble
with \code{n^2} rows is returned.
For a quantitative palette (\verb{_qnt} functions), the palette is discretized;
each color is compared with its neighbors. A data frame with \code{n_step} rows is
returned.
}
\description{
You can evaluate distance for categorial palettes (\verb{_cat} functions), or for
quantitative functions (\verb{_qnt} functions). You can evaluate using Euclidean
distance in a color space (\verb{_euclid} functions), or using a metic
(\verb{_metric} functions).
}
