% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd.R
\name{pth_data_cvd}
\alias{pth_data_cvd}
\alias{pth_data_cvd.default}
\alias{pth_data_cvd.character}
\alias{pth_data_cvd.pth_palette}
\alias{pth_data_cvd.pth_mat}
\title{Color data in CVD context}
\usage{
pth_data_cvd(x, cvd = pth_cvd_grid(), ...)

\method{pth_data_cvd}{default}(x, cvd = pth_cvd_grid(), ...)

\method{pth_data_cvd}{character}(x, cvd = pth_cvd_grid(), transformer = pth_to_cieluv, ...)

\method{pth_data_cvd}{pth_palette}(x, cvd = pth_cvd_grid(), n_point = 11, ...)

\method{pth_data_cvd}{pth_mat}(x, cvd = pth_cvd_grid(), ...)
}
\arguments{
\item{x}{\code{character}, \code{pth_mat}, or \code{pth_palette}: a means to express color}

\item{cvd}{\code{tbl_df} with columns \code{condition}, \code{severity};
see \code{\link[=pth_cvd_grid]{pth_cvd_grid()}}.}

\item{...}{other args, not used.}

\item{transformer}{\code{function} used to transform the colors to new
color space, e.g. \link{pth_to_cieluv}.}

\item{n_point}{\code{integer} number of points from the palette.}
}
\value{
\code{tibble} with columns \code{"luminance"}, \code{"chroma"}, \code{"hue"}, \code{"hex"},
\code{"condition"}, \code{"severity"}
}
\description{
Color data in CVD context
}
