% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-data.R
\name{pth_data_palette}
\alias{pth_data_palette}
\alias{pth_data_palette.default}
\alias{pth_data_palette.pth_palette}
\title{Dataset for palette}
\usage{
pth_data_palette(palette, ...)

\method{pth_data_palette}{default}(palette, ...)

\method{pth_data_palette}{pth_palette}(palette, n = 10, ...)
}
\arguments{
\item{palette}{\code{function} with S3 class \code{pth_palette}}

\item{...}{other arguments (not used).}

\item{n}{\code{numeric} number of colors.}
}
\value{
\code{tibble} with columns \code{luminance}, \code{chroma}
}
\description{
Dataset for palette
}
