% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroma.R
\name{pth_in_gamut}
\alias{pth_in_gamut}
\title{Determine if color is in RGB gamut}
\usage{
pth_in_gamut(color)
}
\arguments{
\item{color}{\code{character} or \code{matrix} with S3 class \code{pth_mat}, a
representation of a color.}
}
\value{
\code{logical} for each color indicates if in-gamut
}
\description{
Determine if color is in RGB gamut
}
\examples{
  pth_in_gamut("#663399")
}
