% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-new.R
\name{pth_layer_control_points}
\alias{pth_layer_control_points}
\alias{pth_layer_control_points.default}
\alias{pth_layer_control_points.data.frame}
\alias{pth_layer_control_points.pth_chroma_trajectory}
\alias{pth_layer_control_points.pth_palette_path}
\title{Layer for control points}
\usage{
pth_layer_control_points(x, ...)

\method{pth_layer_control_points}{default}(x, ...)

\method{pth_layer_control_points}{data.frame}(x, color_light = "grey67", color_dark = "grey33", ...)

\method{pth_layer_control_points}{pth_chroma_trajectory}(x, color_light = "grey67", color_dark = "grey33", ...)

\method{pth_layer_control_points}{pth_palette_path}(x, color_light = "grey67", color_dark = "grey33", ...)
}
\arguments{
\item{x}{\code{data.frame} or \code{function} with S3 class \code{pth_chroma_trajectory} or
\code{pth_palette_path}}

\item{...}{other args (not used)}

\item{color_light, color_dark}{\code{character} R colors to use for symbols
and lines.}
}
\value{
\code{list} of ggplot2 geoms.
}
\description{
Layer for control points
}
