% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-new.R
\name{pth_layer_palette}
\alias{pth_layer_palette}
\title{Layer for palette points}
\usage{
pth_layer_palette(
  palette,
  n_point = 11,
  color_point_inner = "white",
  color_point_outer = "black",
  size_point_inner = 6,
  size_point_outer = size_point_inner + 1,
  n_line = 200,
  color_line = color_point_outer,
  ...
)
}
\arguments{
\item{palette}{\code{function} with S3 class \code{pth_palette}.}

\item{n_point}{\code{numeric} number of points to show.}

\item{color_point_inner, color_point_outer}{\code{character} colors to use
to outline points.}

\item{size_point_inner, size_point_outer}{\code{numeric} size of the points.}

\item{n_line}{\code{numeric} number of segments in the connecting line.}

\item{color_line}{\code{character} \code{character} colors to use
for the connecting line.}

\item{...}{other args (not used)}
}
\value{
\code{list} of ggplot2 geoms.
}
\description{
Layer for palette points
}
