% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamut.R
\name{pth_mat_gamut}
\alias{pth_mat_gamut}
\title{Get colors on sRGB grid}
\usage{
pth_mat_gamut(
  n_point = 5,
  transformer = pth_to_cieluv,
  set = c("edge", "vertex", "surface", "all")
)
}
\arguments{
\item{n_point}{\code{integer} number of points along a dimension}

\item{transformer}{\code{function} used to transform the colors to new
color space, e.g. \link{pth_to_cieluv}.}

\item{set}{\code{character} indicates the geometry to which to limit the points
on the gamut.}
}
\value{
\code{matrix} with S3 class \code{pth_mat}.
}
\description{
Get colors on sRGB grid
}
