% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroma.R
\name{pth_mat_max_chroma}
\alias{pth_mat_max_chroma}
\alias{pth_mat_max_chroma.default}
\alias{pth_mat_max_chroma.pth_surface}
\title{Get colors with maximum chroma}
\usage{
pth_mat_max_chroma(x, ...)

\method{pth_mat_max_chroma}{default}(x, ...)

\method{pth_mat_max_chroma}{pth_surface}(x, n_step = 20, ...)
}
\arguments{
\item{x}{\code{object} to consider.}

\item{...}{other args, not used.}

\item{n_step}{coercible to \code{integer}, number of steps to take in luminance.}
}
\value{
\code{pth_mat}
}
\description{
Get colors with maximum chroma
}
