% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_replace_data.R
\name{pth_mat_replace_data}
\alias{pth_mat_replace_data}
\alias{pth_mat_replace_data.default}
\alias{pth_mat_replace_data.pth_mat}
\title{Replace data in \code{pth_mat}}
\usage{
pth_mat_replace_data(mat, ...)

\method{pth_mat_replace_data}{default}(mat, ...)

\method{pth_mat_replace_data}{pth_mat}(mat, new_data, ...)
}
\arguments{
\item{mat}{\code{double} \code{matrix} with three columns, one row for each color.}

\item{...}{other args (not used)}

\item{new_data}{\code{matrix} with 3 columns, new data for \code{mat}}
}
\value{
Object with same class, attributes as \code{mat}
}
\description{
Replace data in \code{pth_mat}
}
