% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroma.R
\name{pth_max_chroma}
\alias{pth_max_chroma}
\title{Determine maximum chroma}
\usage{
pth_max_chroma(mat)
}
\arguments{
\item{mat}{\code{double} \code{matrix} with three columns, one row for each color.}
}
\value{
\code{double} maximum chroma, one value for each row in \code{mat}.
}
\description{
The chroma is calculated given the color space used in \code{mat}.
}
