% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_color.R
\name{pth_n_color}
\alias{pth_n_color}
\alias{pth_n_color.default}
\alias{pth_n_color.character}
\alias{pth_n_color.pth_hex}
\alias{pth_n_color.pth_mat}
\title{Number of colors}
\usage{
pth_n_color(color, ...)

\method{pth_n_color}{default}(color, ...)

\method{pth_n_color}{character}(color, ...)

\method{pth_n_color}{pth_hex}(color, ...)

\method{pth_n_color}{pth_mat}(color, ...)
}
\arguments{
\item{color}{\code{character} or \code{matrix} with S3 class \code{pth_mat}, a
representation of a color.}

\item{...}{other args (not used)}
}
\value{
\code{integer} number of colors
}
\description{
Number of colors
}
\examples{
  pth_n_color(c("#000000", "#663399", "#FFFFFF"))
}
