% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface.R
\name{pth_new_surface}
\alias{pth_new_surface}
\alias{pth_new_surface.default}
\alias{pth_new_surface.character}
\title{Create luminance-hue surface}
\usage{
pth_new_surface(colors, route = c("short", "long"), n_step = 30, ...)

\method{pth_new_surface}{default}(colors, route = c("short", "long"), n_step = 30, ...)

\method{pth_new_surface}{character}(
  colors,
  route = c("short", "long"),
  n_step = 30,
  transformer = pth_to_cieluv,
  ...
)
}
\arguments{
\item{colors}{\code{character} hex-codes, or matrix with object with
S3 class \code{pth_mat}. Must have one or two elements.}

\item{route}{\code{character}, indicates which direction to take around the
circle; must be either \code{"short"} (default), or \code{"long"}.}

\item{n_step}{(\code{double} coerced to) \code{integer} number of steps to take along
luminance domain to calculate maximum chroma.}

\item{...}{other args, not used}

\item{transformer}{\code{function} used to transform to a color space.
Used if \code{colors} are hex-codes.}
}
\value{
Onject with S3 class \code{pth_new_surface}.
}
\description{
Create luminance-hue surface
}
