% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-join.R
\name{pth_palette_join}
\alias{pth_palette_join}
\alias{pth_palette_join.default}
\alias{pth_palette_join.pth_palette_hex}
\alias{pth_palette_join.pth_palette_path}
\title{Join palette functions}
\usage{
pth_palette_join(palette_low, palette_high, ...)

\method{pth_palette_join}{default}(palette_low, palette_high, ...)

\method{pth_palette_join}{pth_palette_hex}(palette_low, palette_high, ...)

\method{pth_palette_join}{pth_palette_path}(palette_low, palette_high, ...)
}
\arguments{
\item{palette_low, palette_high}{\code{function} with S3 class \code{pth_palette},
palettes to join. \code{palette_low} will be reversed; both palettes will be
rescaled then joined.}

\item{...}{other args (not used)}
}
\value{
\code{function} with S3 class \code{pth_palette},
same as \code{palette_low}, \code{palette_high}.
}
\description{
Use this function to join two sequential palettes into a diverging palette.
}
\examples{
 # both these palettes go from light grey to a darker color
 pal_hex_blue <- pth_new_palette_hex(c("#e2e2e2", "#9cbaee", "#3c79c0"))
 pal_hex_orange <- pth_new_palette_hex(c("#e2e2e2", "#e0af85", "#a66a00"))

 # palette_low (in this case, blue) is reversed,
 # putting the light grey in the middle
 pal_hex_join <- pth_palette_join(pal_hex_blue, pal_hex_orange)

 # recover the original hex codes
 pal_hex_join(seq(0, 1, by = 0.25)) \%>\% pth_to_hex()

}
