% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-rescale.R
\name{pth_palette_rescale_reverse}
\alias{pth_palette_rescale_reverse}
\alias{pth_palette_rescale_euclid}
\alias{pth_palette_rescale_metric}
\title{Rescale palette-function}
\usage{
pth_palette_rescale_reverse(palette)

pth_palette_rescale_euclid(
  palette,
  tolerance = 1e-04,
  non_luminance_weight = 1,
  transformer = identity,
  ...
)

pth_palette_rescale_metric(
  palette,
  tolerance = 1e-04,
  method = c("cie2000", "cie94", "cie1976", "cmc")
)
}
\arguments{
\item{palette}{\code{function} with S3 class \code{pth_palette}.}

\item{tolerance}{\code{numeric} relative tolerance for distance calculations.}

\item{non_luminance_weight}{\code{numeric} used to "discount" the effects of
chroma and hue in the distance calculation.}

\item{transformer}{\code{function} used to transform the colors to the same
color space, e.g. \link{pth_to_cieluv}.}

\item{...}{additional arguments passed on to \code{transformer}.}

\item{method}{\code{character} metric used by \code{farver::compare_colour()}.}
}
\value{
\code{function} with the same S3 classes as \code{palette}
}
\description{
The purpose of these functions is to rescale the input to palette functions.
Inputs to palette functions must be \verb{0 <= x <= 1}. A rescaler maps \code{x} to
\verb{x'} such that \verb{0 <= x' <= 1}; a mapping function will be monotonic.

\code{pth_palette_rescale_reverse()} reverses the input, so that \code{x} maps to
\code{1 - x}.

The other rescaling functions, \code{pth_palette_rescale_euclid()} and
\code{pth_palette_rescale_metric()}, are used so to rescale a palette function
such that a constant change in input, \code{x}, results in a constant change
in color, for a given definition of "change in color".
}
