% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-lumchroma.R
\name{pth_plot_lumchroma}
\alias{pth_plot_lumchroma}
\alias{pth_plot_lumchroma.default}
\alias{pth_plot_lumchroma.character}
\alias{pth_plot_lumchroma.pth_hex}
\alias{pth_plot_lumchroma.pth_mat}
\alias{pth_plot_lumchroma.data.frame}
\title{Plot colors with CVD in luminance-chroma plane}
\usage{
pth_plot_lumchroma(x, ...)

\method{pth_plot_lumchroma}{default}(x, ...)

\method{pth_plot_lumchroma}{character}(
  x,
  cvd = pth_cvd_grid_deupro(),
  transformer = NULL,
  color_point_inner = "white",
  color_point_outer = "black",
  size_point_inner = 6,
  size_point_outer = 7,
  ...
)

\method{pth_plot_lumchroma}{pth_hex}(
  x,
  cvd = pth_cvd_grid_deupro(),
  transformer = NULL,
  color_point_inner = "white",
  color_point_outer = "black",
  size_point_inner = 6,
  size_point_outer = 7,
  ...
)

\method{pth_plot_lumchroma}{pth_mat}(
  x,
  cvd = pth_cvd_grid_deupro(),
  color_point_inner = "white",
  color_point_outer = "black",
  size_point_inner = 6,
  size_point_outer = 7,
  ...
)

\method{pth_plot_lumchroma}{data.frame}(
  x,
  name_color_space = NULL,
  color_point_inner = "white",
  color_point_outer = "black",
  size_point_inner = 6,
  size_point_outer = 7,
  ...
)
}
\arguments{
\item{x}{Object coerced into a CVD data-frame (\code{pth_data_cvd()}). Can be a
hex-code, \code{pth_mat}, or even a data frame.}

\item{...}{other args (not used)}

\item{cvd}{\code{tbl_df} with columns \code{condition}, \code{severity};
see \code{\link[=pth_cvd_grid]{pth_cvd_grid()}}.}

\item{transformer}{\code{function} used to transform the colors to new
color space, e.g. \link{pth_to_cieluv}.}

\item{color_point_inner}{\code{character} colors to use
to outline points.}

\item{color_point_outer}{\code{character} colors to use
to outline points.}

\item{size_point_inner}{\code{numeric} size of the points.}

\item{size_point_outer}{\code{numeric} size of the points.}

\item{name_color_space}{\code{character} name of the color space used.}
}
\value{
Object with S3 class \code{ggplot}.
}
\description{
Oversimplifying things, color-vision deficiency collapses the color-wheel
into a "color-line", such that luminance and chroma are the important
dimensions.
}
