% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-new.R
\name{pth_plot_palette}
\alias{pth_plot_palette}
\alias{pth_plot_palette.default}
\alias{pth_plot_palette.pth_palette_path}
\title{Plot a palette}
\usage{
pth_plot_palette(palette, ...)

\method{pth_plot_palette}{default}(palette, ...)

\method{pth_plot_palette}{pth_palette_path}(
  palette,
  n_point = 11,
  step = 0.5,
  args_control = waiver(),
  args_palette = waiver(),
  ...
)
}
\arguments{
\item{palette}{\code{function} with S3 class \code{pth_palette}.}

\item{...}{other args (not used)}

\item{n_point}{\code{numeric} number of points to show.}

\item{step}{\code{numeric} size of step in luminance and chroma.}

\item{args_control}{\code{list} of additional arguments sent to
\code{\link[=pth_layer_control_points]{pth_layer_control_points()}}, set to \code{NULL} to suppress layer.}

\item{args_palette}{\code{list} of additional arguments sent to
\code{\link[=pth_layer_palette]{pth_layer_palette()}}, set to \code{NULL} to suppress layer.}
}
\value{
\code{tibble} with columns \code{luminance}, \code{chroma}, \code{hue}, \code{hex}
}
\description{
Plot a palette
}
