% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-polar.R
\name{pth_plot_polar}
\alias{pth_plot_polar}
\alias{pth_plot_polar.default}
\alias{pth_plot_polar.data.frame}
\alias{pth_plot_polar.character}
\alias{pth_plot_polar.pth_mat}
\alias{pth_plot_polar.pth_palette}
\title{Polar plot}
\usage{
pth_plot_polar(data, ...)

\method{pth_plot_polar}{default}(data, ...)

\method{pth_plot_polar}{data.frame}(data, ...)

\method{pth_plot_polar}{character}(
  data,
  cvd = pth_cvd_grid_none(),
  transformer = pth_to_cieluv,
  ...
)

\method{pth_plot_polar}{pth_mat}(data, cvd = pth_cvd_grid_none(), ...)

\method{pth_plot_polar}{pth_palette}(data, cvd = pth_cvd_grid_none(), n_point = 11, ...)
}
\arguments{
\item{data}{object coercible to \code{data.frame} with columns
\code{luminance}, \code{chroma}, \code{hue}, \code{hex}.}

\item{...}{other args, not used.}

\item{cvd}{\code{tbl_df} with columns \code{condition}, \code{severity};
see \code{\link[=pth_cvd_grid]{pth_cvd_grid()}}.}

\item{transformer}{\code{function} used to transform the colors to new
color space, e.g. \link{pth_to_cieluv}.}

\item{n_point}{\code{integer} number of points from the palette.}
}
\value{
Object with S3 classes \code{"gg"}, \code{"ggplot"}; i.e. a ggplot
}
\description{
Polar plot
}
