% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface-data.R
\name{pth_surface_data}
\alias{pth_surface_data}
\alias{pth_surface_data.default}
\alias{pth_surface_data.pth_surface}
\title{Dataset for surfaces}
\usage{
pth_surface_data(x, ...)

\method{pth_surface_data}{default}(x, ...)

\method{pth_surface_data}{pth_surface}(x, step = 0.5, ...)
}
\arguments{
\item{x}{\code{object} with S3 class \code{pth_surface}.}

\item{...}{other args, not used.}

\item{step}{\code{numeric} size of step in luminance and chroma.}
}
\value{
\code{tibble} with columns \code{luminance}, \code{chroma}, \code{hue}, \code{hex}.
}
\description{
Dataset for surfaces
}
