% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.R
\name{pth_to_cartesian}
\alias{pth_to_cartesian}
\alias{pth_to_polar}
\title{Convert between Cartesian and polar coordinates}
\usage{
pth_to_cartesian(mat_polar, chroma_min = 1e-04)

pth_to_polar(mat_cartesian)
}
\arguments{
\item{mat_polar}{\code{matrix} with columns for luminance, chroma, hue.}

\item{chroma_min}{\code{numeric} "trick" used to preserve the hue at zero chroma,
using insignificant decimal places.}

\item{mat_cartesian}{\code{matrix} with columns for luminance,
blue-yellow, green-red.}
}
\value{
\code{matrix} with same dimension as the input matrix.
}
\description{
Convert between Cartesian and polar coordinates
}
\examples{
  polar <- matrix(c(50, 0, 45), ncol = 3)
  cart <- pth_to_cartesian(polar)
  polar_new <- pth_to_polar(cart)
  # notice the chroma value has an artificial floor
  # using an imperceptible (color) difference
  print(polar)
  print(cart)
  print(polar_new)
}
