% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cielab.R
\name{pth_to_cielab}
\alias{pth_to_cielab}
\alias{pth_new_cielab}
\title{Convert/coerce to CIELAB matrix}
\usage{
pth_to_cielab(color, whitepoint = whitepoints_cie1931("D65"))

pth_new_cielab(mat, whitepoint = whitepoints_cie1931("D65"))
}
\arguments{
\item{color}{\code{character} or \code{matrix} with S3 class \code{pth_mat}, a
representation of a color.}

\item{whitepoint}{\code{double} \code{array} with one dimension, length three;
describes the whitepoint reference for the color space.}

\item{mat}{\code{double} \code{matrix} with three columns, one row for each color.}
}
\value{
\code{double} \code{matrix} with S3 classes \code{pth_cielab} and \code{pth_mat},
with three columns, one row for each color.
}
\description{
Convert/coerce to CIELAB matrix
}
\examples{
  pth_to_cielab("#663399")
  pth_new_cielab(matrix(c(32.9, 42.9, -47.1), ncol = 3))
}
