% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cieluv.R
\name{pth_to_cieluv}
\alias{pth_to_cieluv}
\alias{pth_new_cieluv}
\title{Convert/coerce to CIELUV matrix}
\usage{
pth_to_cieluv(color, whitepoint = whitepoints_cie1931("D65"))

pth_new_cieluv(mat, whitepoint = whitepoints_cie1931("D65"))
}
\arguments{
\item{color}{\code{character} or \code{matrix} with S3 class \code{pth_mat}, a
representation of a color.}

\item{whitepoint}{\code{double} \code{array} with one dimension, length three;
describes the whitepoint reference for the color space.}

\item{mat}{\code{double} \code{matrix} with three columns, one row for each color.}
}
\value{
\code{double} \code{matrix} with S3 classes \code{pth_cieluv} and \code{pth_mat},
with three columns, one row for each color.
}
\description{
Convert/coerce to CIELUV matrix
}
\examples{
  pth_to_cieluv("#663399")
  pth_new_cieluv(matrix(c(32.9, 13.0, -67.8), ncol = 3))
}
