% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{pth_to_hex}
\alias{pth_to_hex}
\alias{pth_to_hex.default}
\alias{pth_to_hex.pth_hex}
\alias{pth_to_hex.character}
\title{Convert to hex code}
\usage{
pth_to_hex(color, ...)

pth_to_hex.default(color, ...)

pth_to_hex.pth_hex(color, ...)

pth_to_hex.character(color, ...)
}
\arguments{
\item{color}{\code{character} or \code{matrix} with S3 class \code{pth_mat}, a
representation of a color.}

\item{..., }{other arguments as may be needed.}
}
\value{
\code{character} with S3 class \code{pth_hex}
}
\description{
Convert to hex code
}
\examples{
  pth_to_hex("#663399")
  pth_to_hex(c("#221133", "#442266", "#663399"))
}
