% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jzazbz100.R
\name{pth_to_jzazbz100}
\alias{pth_to_jzazbz100}
\alias{pth_new_jzazbz100}
\title{Convert/coerce to JzAzBz-100 matrix}
\usage{
pth_to_jzazbz100(color, whitepoint = whitepoints_cie1931("D65"))

pth_new_jzazbz100(mat, whitepoint = whitepoints_cie1931("D65"))
}
\arguments{
\item{color}{\code{character} or \code{matrix} with S3 class \code{pth_mat}, a
representation of a color.}

\item{whitepoint}{\code{double} \code{array} with one dimension, length three;
describes the whitepoint reference for the color space.}

\item{mat}{\code{double} \code{matrix} with three columns, one row for each color.}
}
\value{
\code{double} \code{matrix} with S3 classes \code{pth_jzazbz100} and \code{pth_mat},
with three columns, one row for each color.
}
\description{
The idea for the scaling came from
https://github.com/nschloe/colorio/issues/41#issuecomment-519983981.
}
\examples{
  pth_to_jzazbz100("#663399")
  pth_new_jzazbz100(matrix(c(36.3, 15,3, -42.6), ncol = 3))
}
