% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam02ucs.R, R/cam16ucs.R, R/cielab.R,
%   R/cieluv.R, R/jzazbz100.R, R/transformer.R
\name{pth_transformer.pth_cam02ucs}
\alias{pth_transformer.pth_cam02ucs}
\alias{pth_transformer.pth_cam16ucs}
\alias{pth_transformer.pth_cielab}
\alias{pth_transformer.pth_cieluv}
\alias{pth_transformer.pth_jzazbz100}
\alias{pth_transformer}
\title{Extract transformer function}
\usage{
\method{pth_transformer}{pth_cam02ucs}(mat, ...)

\method{pth_transformer}{pth_cam16ucs}(mat, ...)

\method{pth_transformer}{pth_cielab}(mat, ...)

\method{pth_transformer}{pth_cieluv}(mat, ...)

\method{pth_transformer}{pth_jzazbz100}(mat, ...)

pth_transformer(mat, ...)
}
\arguments{
\item{mat}{matrix with object with S3 class \code{pth_mat}}

\item{...}{other args, not used}
}
\value{
\code{function} that encapsulates a transformer, i.e. \code{pth_to_cieluv}
}
\description{
Extract transformer function
}
