% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_pal_luv.R
\name{rescale_pal_luv}
\alias{rescale_pal_luv}
\title{Rescale palette}
\usage{
rescale_pal_luv(pal_luv, rescaler)
}
\arguments{
\item{pal_luv}{\code{function} with S3 class \code{"cpath_pal_luv"}}

\item{rescaler}{\code{function} with S3 class \code{"cpath_rescaler"}}
}
\value{
\code{function} with S3 class \code{"cpath_pal_luv"}
}
\description{
Rescale a palette function using a rescaler function.
}
\examples{
  # create original palette
  pal_blues <- palette_bezier(mat_luv_blues)

  # create rescaler
  rsc_reverse <- rescaler_x(c(1, 0))

  # create reversed palette
  pal_blues_reverse <- rescale_pal_luv(pal_blues, rsc_reverse)

  pal_blues(0)
  pal_blues_reverse(1)

}
