% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale-palette.R
\name{rescale_palette}
\alias{rescale_palette}
\title{Rescale palette}
\usage{
rescale_palette(palette, rescaler)
}
\arguments{
\item{palette}{\code{function} with S3 class \code{"cpath_palette_luv"}}

\item{rescaler}{\code{function} with S3 class \code{"cpath_rescaler"}}
}
\value{
\code{function} with S3 class \code{"cpath_palette_luv"}
}
\description{
Rescale a palette function using a rescaler function.
}
\examples{
  # create original palette
  pal_blues <- palette_bezier(mat_luv_blues)

  # create rescaler
  rsc_reverse <- rescaler_linear_input(c(1, 0))

  # create reversed palette
  pal_blues_reverse <- rescale_palette(pal_blues, rsc_reverse)

  pal_blues(0)
  pal_blues_reverse(1)

}
