% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaler.R
\name{rescaler_bezier}
\alias{rescaler_bezier}
\title{Bézier rescaler}
\usage{
rescaler_bezier(mat_luv, n = 21)
}
\arguments{
\item{mat_luv}{\code{matrix} of LUV coordinates, control-points for Bézier spline.}

\item{n}{\code{numeric} if \code{rescale_path}, number of equally-spaced Bézier points to calculate.}
}
\value{
A function with S3 class \code{cpath_rescaler}.
}
\description{
Bézier rescaler
}
\keyword{internal}
