% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaler.R
\name{rescaler_dist}
\alias{rescaler_dist}
\title{Perceptual-distance rescaler}
\usage{
rescaler_dist(pal_luv, tolerance = 1e-04, method = "cie2000", ...)
}
\arguments{
\item{pal_luv}{\code{function} with S3 class \code{"cpath_pal_luv"}}

\item{tolerance}{\code{numeric} convergence tolerance for distance calculation}

\item{method}{\code{character} method to use for distance, one of
\code{"cie2000"}, \code{"euclidean"}, \code{"cie1976"}, \code{"cie94"}, or \code{"cmc"}.}

\item{...}{other arguments sent to \code{\link[farver:compare_colour]{farver::compare_colour()}}.}
}
\value{
A function with S3 class \code{cpath_rescaler}.
}
\description{
Perceptual-distance rescaler
}
\keyword{internal}
