% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_hl.R
\name{surface_hl}
\alias{surface_hl}
\title{Create HL surface-function}
\usage{
surface_hl(range_hue, input_lum = c(0, 100))
}
\arguments{
\item{range_hue}{\code{numeric} vector, length 1 or 2, specifies the
hue at the luminance values specified in \code{input_lum}. If only one
hue is specified, this will return a function for constant hue.}

\item{input_lum}{\code{numeric} vector, length 2, specifies the
luminance values corresponding to \code{range_hue}.}
}
\value{
\code{function} that takes luminance values, returns hue values.
}
\description{
Creates a function that defines hue in terms of luminance.
}
\details{
The \code{range_hue} you specify need not be between 0 and 360 - in fact you
might not want them to be if you want to define a surface that crosses the
hue branch-cut. The surface function will return hue values between 0 and 360.
}
\examples{
  # single hue
  sfc_blues_single <- surface_hl(250)
  sfc_blues_single(c(0, 50, 100))

  # multi hue
  sfc_blues_multi <- surface_hl(c(240, 260))
  sfc_blues_multi(c(0, 50, 100))

}
