% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitepoint.R
\name{whitepoint_cie1931}
\alias{whitepoint_cie1931}
\title{Access colorio whitepoints}
\usage{
whitepoint_cie1931(name)
}
\arguments{
\item{name}{\code{character} a legal name of a whitepoint:
\code{"A"}, \code{"C"}, \code{"D50"}, \code{"D55"}, \code{"D65"}, \code{"D75"}}
}
\value{
\code{numeric} \code{array} with one dimension of length 3.
}
\description{
Access colorio whitepoints
}
\examples{
  # not run because it requires Python
  if (FALSE) {
    whitepoint_cie1931("D65")
  }
}
