% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroma.R
\name{x_gamut_chroma}
\alias{x_gamut_chroma}
\title{Determine gamut-distance for a chroma, given a color}
\usage{
x_gamut_chroma(chroma, mat)
}
\arguments{
\item{chroma}{\code{numeric} value for chroma, expressed using color space
for \code{mat}.}
}
\value{
\code{double} indicating "distance" from gamut surface,
negative indicates inside gamut.
}
\description{
Determine gamut-distance for a chroma, given a color
}
\keyword{internal}
