% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_pal_hex.R
\name{as_pal_hex}
\alias{as_pal_hex}
\title{Coerce to hex-code palette-function}
\usage{
as_pal_hex(pal_luv)
}
\arguments{
\item{pal_luv}{\code{function} with S3 class \code{"cpath_palette_luv"}}
}
\value{
\code{function} given a numeric input \verb{(0 <= x <= 1)}, return hex-codes
}
\description{
Most of the other
}
\examples{
  # create LUV palette-function
  pal_blues <- palette_bezier(mat_luv_blues)

  # create hex palette-function
  pal_blues_hex <- as_pal_hex(pal_blues)

  # evaluate
  pal_blues_hex(seq(0, 1, by = 0.2))

}
