% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luv.R
\name{luv}
\alias{luv}
\title{Convert to \code{luv} matrix from \code{hcl} data frame}
\usage{
luv(df_hcl)
}
\arguments{
\item{df_hcl}{\code{data.frame} with variables \code{h}, \code{c}, \code{l}}
}
\value{
\code{matrix} with columns, \code{l}, \code{u}, \code{v} and same number
of rows as \code{df_hcl}
}
\description{
It is thought to specify a colorpath using HCL coordinates;
it seems reasonable to compile these coordinates using a
data frame.
}
\examples{
  df_hcl <- data.frame(h = c(0, 30), c = c(50, 50), l = c(50, 60))
  luv(df_hcl)
}
