% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal_luv_rescale.R
\name{pal_luv_rescale}
\alias{pal_luv_rescale}
\title{Rescale LUV palette-function}
\usage{
pal_luv_rescale(pal_luv, rescaler)
}
\arguments{
\item{pal_luv}{\code{function} with S3 class \code{"cpath_pal_luv"}}

\item{rescaler}{\code{function} with S3 class \code{"cpath_rescaler"}}
}
\value{
\code{function} with S3 class \code{"cpath_pal_luv"}
}
\description{
Rescale LUV palette function using a rescaler function.
}
\examples{
  # create original palette
  pal_blues <- pal_luv_bezier(mat_luv_blues)

  # create rescaler
  rsc_reverse <- rescaler_x(c(1, 0))

  # create reversed palette
  pal_blues_reverse <- pal_luv_rescale(pal_blues, rsc_reverse)

  pal_blues(0)
  pal_blues_reverse(1)

}
