% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-bezier.R
\name{palette_bezier}
\alias{palette_bezier}
\title{Create palette-function using Bézier spline}
\usage{
palette_bezier(mat_luv)
}
\arguments{
\item{mat_luv}{\code{matrix} of LUV coordinates, control-points for Bézier spline.}
}
\value{
A function with S3 class \code{cpath_palette}.
}
\description{
Thus takes a \code{matrix} of LUV coordinates and returns a palette-function
(that returns LUV values) based on a Bézier spline. You can get an \code{LUV}
matrix using the function \code{\link[=as_mat_luv]{as_mat_luv()}}.
}
\examples{
 # create palette-function
 pb <- palette_bezier(mat_luv_blues)

 # evaluate
 pb(c(0, 0.5, 1))

}
