% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-bezier.R
\name{palette_bezier}
\alias{palette_bezier}
\title{Create palette-function using Bézier spline}
\usage{
palette_bezier(luv)
}
\arguments{
\item{luv}{\code{matrix} of LUV coordinates, control-points for Bézier spline.}
}
\value{
A function with S3 class \code{cpath_palette}.
}
\description{
Thus takes a \code{matrix} of LUV coordinates and returns a palette-function
(that returns LUV values) based on a Bézier spline. You can get an \code{LUV}
matrix using the function \code{\link[=luv]{luv()}}.
}
\examples{
 # matrix of LUV values
 mat_luv <- matrix(
   c(
     20,  50, 80, # l
     0,  -46,  0, # u
     0,   56,  0  # v
   ),
   ncol = 3L,
   byrow = FALSE,
   dimnames = list(NULL, c("l", "u", "v"))
 )

 # create palette-function
 pb <- palette_bezier(mat_luv)

 # evaluate
 pb(c(0, 0.5, 1))

}
