% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd.R
\name{pth_cvd_grid}
\alias{pth_cvd_grid}
\alias{pth_cvd_grid_full}
\title{Color-vision deficiency grid}
\usage{
pth_cvd_grid(condition = c("none", "deutan", "protan", "tritan"), severity = 1)

pth_cvd_grid_full(
  condition = c("deutan", "protan", "tritan"),
  severity = c(0, 0.33, 0.67, 1)
)
}
\arguments{
\item{condition}{\code{character} one or more color-vision deficiency conditions;
legal values: \code{"none"}, \code{"deutan"}, \code{"protan"}, \code{"tritan"}.}

\item{severity}{\code{double} one or more values of severity for a given
condition; legal values from 0 to 1.}
}
\value{
\code{tbl_df} with columns \code{cvd} (factor), \code{severity} (double)
}
\description{
Helper function to generate grid for color-vision deficiency (CVD)
possibilities.
}
