% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-rescale.R
\name{pth_palette_rescale_reverse}
\alias{pth_palette_rescale_reverse}
\alias{pth_palette_rescale_euclid}
\alias{pth_palette_rescale_metric}
\title{Rescale palette-function}
\usage{
pth_palette_rescale_reverse(palette)

pth_palette_rescale_euclid(
  palette,
  tolerance = 1e-04,
  non_luminance_weight = 1,
  transformer = identity,
  ...
)

pth_palette_rescale_metric(
  palette,
  tolerance = 1e-04,
  method = c("cie2000", "cie94", "cie1976", "cmc")
)
}
\arguments{
\item{palette}{\code{function} with S3 class \code{pth_palette}.}

\item{tolerance}{\code{numeric} relative tolerance for distance calculations.}

\item{non_luminance_weight}{\code{numeric} used to "discount" the effects of
chroma and hue in the distance calculation.}

\item{transformer}{\code{function} used to transform the colors to the same
color space, e.g. \link{pth_to_cieluv}.}

\item{...}{additional arguments passed on to \code{transformer}.}

\item{method}{\code{character} metric used by \code{farver::compare_colour()}.}
}
\value{
\code{function} with the same S3 classes as \code{palette}
}
\description{
Rescale palette-function
}
