% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-polar.R
\name{pth_plot_polar}
\alias{pth_plot_polar}
\alias{pth_plot_polar.default}
\alias{pth_plot_polar.data.frame}
\alias{pth_plot_polar.character}
\alias{pth_plot_polar.pth_mat}
\alias{pth_plot_polar.pth_palette}
\title{Polar plot}
\usage{
pth_plot_polar(data, ...)

\method{pth_plot_polar}{default}(data, ...)

\method{pth_plot_polar}{data.frame}(data, ...)

\method{pth_plot_polar}{character}(
  data,
  cvd = pth_cvd_grid_none(),
  transformer = pth_to_cieluv,
  ...
)

\method{pth_plot_polar}{pth_mat}(data, cvd = pth_cvd_grid_none(), ...)

\method{pth_plot_polar}{pth_palette}(data, cvd = pth_cvd_grid_none(), n_point = 11, ...)
}
\arguments{
\item{data}{object coercible to \code{data.frame} with columns
\code{luminance}, \code{chroma}, \code{hue}, \code{hex}.}

\item{...}{other args, not used.}

\item{cvd}{\code{tbl_df} with columns \code{condition}, \code{severity};
see \code{\link[=pth_cvd_grid]{pth_cvd_grid()}}.}

\item{transformer}{\code{function} used to transform the colors to new
color space, e.g. \link{pth_to_cieluv}.}

\item{n_point}{\code{integer} number of points from the palette.}
}
\value{
Object with S3 classes \code{"gg"}, \code{"ggplot"}; i.e. a ggplot.
}
\description{
This can be useful in identifying issues arising from color-vision
deficiency. You can provide a \code{data.frame} that has columns
\code{luminance}, \code{chroma}, \code{hue}, \code{hex} - or you can provide an object
that can be coerced into that form using \code{\link[=pth_data_cvd]{pth_data_cvd()}}: a \code{pth_mat},
\code{pth_palette}, or hex-code \code{character}.
}
\details{
If you provide the \code{cvd} argument anything other than the simplest case,
you will have to add faceting. You will have to do this yourself;
the operative columns are \code{condition} and \code{severity}.
}
