% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaler-new.R
\name{pth_rescaler_reverse}
\alias{pth_rescaler_reverse}
\alias{pth_rescaler_euclid}
\alias{pth_rescaler_metric}
\title{Create rescaler function}
\usage{
pth_rescaler_reverse()

pth_rescaler_euclid(palette, tolerance, transformer, non_luminance_weight, ...)

pth_rescaler_metric(palette, tolerance, method)
}
\arguments{
\item{palette}{\code{function} with S3 class \code{pth_palette}.}

\item{tolerance}{\code{numeric} relative tolerance for distance calculations.}

\item{transformer}{\code{function} used to transform the colors to the same
color space, e.g. \link{pth_to_cieluv}.}

\item{non_luminance_weight}{\code{numeric} used to "discount" the effects of
chroma and hue in the distance calculation.}

\item{...}{additional arguments passed on to \code{transformer}.}

\item{method}{\code{character} metric used by \code{farver::compare_colour()}.}
}
\value{
\code{function}, for each input (\verb{0 <= x <= 1}) returns a
value (\verb{0 <= y <= 1}). The function shall be monotonic.
}
\description{
A rescaler function is used to rescale the input to a palette function.

\describe{
\item{\code{pth_rescaler_reverse()}}{Simply reverses the input.}
\item{\code{pth_rescaler_euclid()}}{Uses \code{palette} and \code{pth_distance_euclid()}
to rescale the input so that the distance-change is constant.}
\item{\code{pth_rescaler_metric()}}{Uses \code{palette} and \code{pth_distance_metric()}
to rescale the input so that the distance-change is constant.}
}
}
\keyword{internal}
